requirements.txt
alembic==1.8.0
amqp==5.0.9
APScheduler==3.9.1
asgiref==3.4.1
asn1crypto==1.5.1
atomicwrites==1.4.0
attrs==21.2.0
billiard==3.6.4.0
cachetools==4.2.4
celery==5.0.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==3.2.14
django-apscheduler==0.6.2
django-cors-headers==3.5.0
django-crontab==0.7.1
django-debug-toolbar==3.2.4
django-environ==0.4.5
django-extensions==3.1.5
django-filter==2.4.0
django-import-export==2.7.1
django-rest-swagger==2.2.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.13.2
drf-yasg==1.20.0
et-xmlfile==1.1.0
Faker==5.8.0
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
google-api-core==1.31.5
google-api-python-client==1.12.8
google-auth==1.30.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==2.3.1
google-cloud-secret-manager==2.11.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
greenlet==1.1.2
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
gunicorn==20.0.4
httplib2==0.20.2
idna==3.3
importlib==1.0.4
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.0.3
jsonschema==4.3.2
kombu==5.2.2
Mako==1.2.1
MarkupPy==1.14
MarkupSafe==2.1.1
MouseInfo==0.1.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.1.1
odfpy==1.4.1
openapi-codec==1.3.2
opencv-python==4.5.5.64
openpyxl==3.0.6
packaging==21.3
pg8000==1.29.1
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.5.1
pluggy==1.0.0
portalocker==2.5.1
prompt-toolkit==3.0.24
proto-plus==1.20.6
protobuf==3.19.1
psycopg2==2.9.2
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.3.0
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyparsing==3.0.6
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.28
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.15.0
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
redis==3.5.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scramp==1.4.1
sib-api-v3-sdk==7.5.0
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
tablib==3.1.0
text-unidecode==1.3
toml==0.10.2
torch==1.11.0
tqdm==4.63.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
uritemplate==3.0.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.13.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.0.3
xlrd==2.0.1
xlwt==1.3.0
